Move device and data models to schema package
