+++ src/swc/aeon/schema/video.py
@@ -4,7 +4,7 @@ from typing import Literal
 
 from pydantic import Field
 
-from swc.aeon.rigs.base import Device
+from swc.aeon.schema.base import Device
 
 
 class SpinnakerCamera(Device):
